refactor(http): migrate HttpError + HttpResult to dcc-mcp-http-types (#852)
Second migration in the #852 chain. Moves the HTTP error taxonomy into
dcc-mcp-http-types so external Rust consumers — REST clients,
integration tests, gateway code that maps HTTP errors into the
cross-crate DccMcpError taxonomy (#488) — can pull just the error
surface without the full HTTP server crate.

Migrated:
- HttpError enum (12 variants: AlreadyRunning, NotRunning, BindFailed,
  SessionNotFound, InvalidSessionId, Json, ExecutorClosed,
  QueueOverloaded, Dispatch, Timeout, RateLimit, Internal).
- HttpResult<T> alias.
- HttpError::status_code() — HTTP status code mapping (404/400/429/
  409/504/503/500).
- From<HttpError> for DccMcpError — the #488 taxonomy bubble.
- All 5 existing tests + 1 new regression guard
  (status_codes_match_documented_contract) so the #488 mapping cannot
  silently drift.

dcc-mcp-http side:
- crates/dcc-mcp-http/src/error.rs is now a re-export shim so the
  historical paths (crate::error::HttpError, crate::error::HttpResult,
  dcc_mcp_http::HttpError) keep compiling unchanged.

dcc-mcp-http-types Cargo.toml additions:
- thiserror = workspace (HttpError uses #[derive(Error)]).
- dcc-mcp-models = path internal dep (DccMcpError is the cross-crate
  error taxonomy; dcc-mcp-models is itself a pure-types crate so the
  dependency direction stays inward).

Documentation in error.rs spells out each variant's responsibility and
the contract for adding a new variant (only when DccMcpError genuinely
cannot carry the caller's branching needs) so the taxonomy does not
sprawl.

Verification:
- cargo check --workspace clean
- cargo test -p dcc-mcp-http-types: 18 passed (12 prior + 6 new error)
- cargo test -p dcc-mcp-http --lib: 225 passed
- cargo clippy -p dcc-mcp-http-types --all-targets -- -D warnings clean
- cargo clippy -p dcc-mcp-http --no-deps -- -D warnings clean
- cargo fmt --all -- --check clean

Refs: #852, builds on #895 (#852 part 1). Part of #848 epic.
